# Improve load/save performance