--- a/docx4j-core/src/main/java/org/docx4j/openpackaging/io3/stores/ZipPartStore.java
+++ b/docx4j-core/src/main/java/org/docx4j/openpackaging/io3/stores/ZipPartStore.java
@@ -21,15 +21,7 @@ package org.docx4j.openpackaging.io3.stores;
 
 import static java.lang.Math.toIntExact;
 
-import java.io.BufferedInputStream;
-import java.io.BufferedOutputStream;
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
-import java.io.Serializable;
+import java.io.*;
 import java.nio.ByteBuffer;
 import java.util.Enumeration;
 import java.util.HashMap;
@@ -50,15 +42,8 @@ import org.docx4j.openpackaging.contenttype.ContentTypeManager;
 import org.docx4j.openpackaging.exceptions.Docx4JException;
 import org.docx4j.openpackaging.exceptions.Docx4JRuntimeException;
 import org.docx4j.openpackaging.exceptions.PartTooLargeException;
-import org.docx4j.openpackaging.parts.CustomXmlDataStoragePart;
-import org.docx4j.openpackaging.parts.JaxbXmlPart;
-import org.docx4j.openpackaging.parts.Part;
-import org.docx4j.openpackaging.parts.PartName;
-import org.docx4j.openpackaging.parts.XmlPart;
-import org.docx4j.openpackaging.parts.WordprocessingML.BinaryPart;
-import org.docx4j.openpackaging.parts.WordprocessingML.EmbeddedPackagePart;
-import org.docx4j.openpackaging.parts.WordprocessingML.ImagePngPart;
-import org.docx4j.openpackaging.parts.WordprocessingML.OleObjectBinaryPart;
+import org.docx4j.openpackaging.parts.*;
+import org.docx4j.openpackaging.parts.WordprocessingML.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.w3c.dom.Document;
@@ -187,55 +172,16 @@ public class ZipPartStore implements PartStore {
 		return (partByteArrays.get(partName) !=null );
 	}
 
-	private byte[] getBytesFromInputStream(InputStream is, long size)
-			throws Exception {
-
+    private byte[] getBytesFromInputStream(InputStream is, long size) throws Exception {
 		if (size == -1) {
-		
-			log.debug("entry.getSize() -1");
-			
-			BufferedInputStream bufIn = new BufferedInputStream(is);
-			ByteArrayOutputStream baos = new ByteArrayOutputStream();
-			BufferedOutputStream bos = new BufferedOutputStream(baos);
-			int c = bufIn.read();
-			while (c != -1) {
-				bos.write(c);
-				c = bufIn.read();
-			}
-			bos.flush();
-			baos.flush();
-			//bufIn.close(); //don't do that, since it closes the ZipInputStream after we've read an entry!
-			bos.close();
-			return baos.toByteArray();
-			
+            log.debug("entry.getSize() -1");
+            return is.readAllBytes();
         } else {
+            log.debug("entry.getSize()={}", +size);
+            return is.readNBytes(toIntExact(size));
+        }
+    }
 
-        	if (log.isDebugEnabled()) {
-        		log.info("entry.getSize()=" + size );
-        	}
-        	
-        	int sizeInt = toIntExact(size);
-            byte[] targetArray = new byte[sizeInt];
-            int read = is.read(targetArray);
-            int offset = read;
-            while (read != -1  && (sizeInt-offset!=0)) {
-               read = is.read(targetArray,offset, sizeInt-offset);
-               offset += read;
-            }
-            return targetArray;
-         }			
-			
-	}
-
-//	private static InputStream getInputStreamFromZippedPart(HashMap<String, ByteArray> partByteArrays,
-//			String partName) throws IOException {
-//
-//        ByteArray bytes = partByteArrays.get(partName);
-//        if (bytes == null) throw new IOException("part '" + partName + "' not found");
-//		return bytes.getInputStream();
-//	}
-
-//	protected InputStream getInputStreamFromZippedPart(String partName) throws IOException {
 	public InputStream loadPart(String partName) throws Docx4JException {
 
         ByteArray bytes = partByteArrays.get(partName);
@@ -491,9 +437,7 @@ public class ZipPartStore implements PartStore {
 	        }
 			
 	        // Add ZIP entry to output stream.
-			if (part instanceof OleObjectBinaryPart 
-					|| part instanceof EmbeddedPackagePart
-					|| part instanceof ImagePngPart) {
+			if (part instanceof OleObjectBinaryPart || ! shouldCompress(part)) {
 				// Workaround: Powerpoint 2010 (32-bit) can't play eg WMV if it is compressed!
 				// (though 64-bit version is fine)
 				
@@ -522,9 +466,19 @@ public class ZipPartStore implements PartStore {
 			throw new Docx4JException("Failed to put binary part", e);
 		}
 
-		log.debug( "success writing part: " + resolvedPartUri);
-
+		log.debug("success writing part: " + resolvedPartUri);
 	}
+
+    private static boolean shouldCompress(Part part) {
+        if (part instanceof EmbeddedPackagePart || part instanceof ImagePngPart || part instanceof ImageJpegPart || part instanceof ImageGifPart)
+            return false;
+        if (part instanceof AlternativeFormatInputPart) {
+            AltChunkType type = ((AlternativeFormatInputPart) part).getAltChunkType();
+            return type != AltChunkType.WordprocessingML && type != AltChunkType.OfficeWordTemplate &&
+                   type != AltChunkType.OfficeWordMacroEnabled && type != AltChunkType.OfficeWordMacroEnabledTemplate;
+        }
+        return true;
+    }
 
 	public void finishSave() throws Docx4JException {