This MR improves the performance of loading and saving files.

### getBytesFromInputStream 
For ZipEntries of unkown size (most entries for files generated by docx4j), ZipPartStore reads each byte individually. This MR changes this to reading byte arrays. 

| Load Performance          | Original load time | MR load time |
|---------------------------|--------------------|--------------|
| File last saved by Word   | 75 ms              | 75 ms        |
| File last saved by docx4j | 250 ms             | 75 ms        |


### canCompress
Mark more file types as incompressible, as zipping and unzipping word or jpeg files is slow and provides negligible benefits.